Change image tunes format
